Final Stage 1 pass: disconnect log level, integer stale setting
Normal (shutdown) disconnects log at INFO so the measurement log shows only
real connection losses as warnings; stale_after stays the configured integer
in status; document clock-step-back behavior of the recorder clamp.

+++ backend/pompa/ingest.py
@@ -45,8 +45,8 @@ class SourceState:
 
 
 class Ingest:
-    def __init__(self, stale_after: float):
-        self.stale_after = float(stale_after)
+    def __init__(self, stale_after: int):
+        self.stale_after = stale_after
         self.sources: dict[str, SourceState] = {
             topic: SourceState(metric, source) for topic, (metric, source) in SOURCE_BY_TOPIC.items()
         }
